refactor(phase4): extract marketplace module
- Created marketplace module with 6 components:
  - GitHubClient: Handles GitHub API interactions with caching and rate limiting
  - MarketplaceBrowser: Browse personas by category
  - MarketplaceSearch: Search for personas in marketplace
  - PersonaDetails: Get detailed persona information
  - PersonaInstaller: Install personas from marketplace
  - PersonaSubmitter: Generate submission issues for GitHub
- Removed ~380 lines from index.ts (now 2,615 lines, down from 3,000)
- Updated imports and class initialization
- Fixed type exports (ErrorCode in mcp.ts)
- All tests pass, build successful
