+++ docs/development/REFACTORING_PROGRESS.md
@@ -9,8 +9,8 @@
 - **Phase 1**: ✅ COMPLETE - Foundation structure created
 - **Phase 2**: ✅ COMPLETE - Core utilities extracted
 - **Phase 3**: ✅ COMPLETE - Persona module extracted
-- **Phase 4**: 🔄 IN PROGRESS - Marketplace module (next task)
-- **Phase 5**: ⏳ PENDING - Update module
+- **Phase 4**: ✅ COMPLETE - Marketplace module extracted
+- **Phase 5**: ⏳ PENDING - Update module (next task)
 - **Phase 6**: ⏳ PENDING - Server refactor
 - **Phase 7**: ⏳ PENDING - Cleanup
 
@@ -18,25 +18,27 @@
 1. `e3f2d36` - refactor(phase1): create module structure and extract types/constants
 2. `b76d978` - refactor(phase2): extract core utilities to separate modules
 3. `584f226` - refactor(phase3): extract persona module
+4. [pending commit] - refactor(phase4): extract marketplace module
 
 ## Lines Extracted
-- Phase 1: Types and constants
+- Phase 1: Types and constants  
 - Phase 2: ~400 lines (utilities)
 - Phase 3: ~630 lines (persona logic)
-- **Total**: ~1,030 lines removed from index.ts
+- Phase 4: ~380 lines (marketplace functions)
+- **Total**: ~1,410 lines removed from index.ts (2,615 lines remaining)
 
-## Next Steps for Phase 4
-Extract marketplace functions from index.ts:
-1. Create `src/marketplace/MarketplaceClient.ts`
-2. Create `src/marketplace/MarketplaceSearch.ts`
-3. Create `src/marketplace/PersonaInstaller.ts`
+## Next Steps for Phase 5
+Extract update/version management functions from index.ts:
+1. Create `src/update/UpdateManager.ts`
+2. Create `src/update/BackupManager.ts`  
+3. Create `src/update/DependencyChecker.ts`
 
-Functions to move (approximate line numbers in current index.ts):
-- `browseMarketplace()` (~979-1065)
-- `searchMarketplace()` (~1067-1141)
-- `getMarketplacePersona()` (~1143-1220)
-- `installPersona()` (~1222-1309)
-- `submitPersona()` (~1311-1379)
+Functions to move:
+- Version checking and comparison logic
+- Backup creation and management
+- Dependency validation
+- Update installation logic
+- Rollback functionality
 
 ## Key Files to Check
 - `src/index.ts` - Still contains remaining code
