+++ src/marketplace/index.ts
@@ -2,7 +2,9 @@
  * Marketplace module exports
  */
 
-// Export marketplace functionality when ready
-// export * from './MarketplaceClient.js';
-// export * from './MarketplaceSearch.js';
-// export * from './PersonaInstaller.js';
+export * from './GitHubClient.js';
+export * from './MarketplaceBrowser.js';
+export * from './MarketplaceSearch.js';
+export * from './PersonaDetails.js';
+export * from './PersonaInstaller.js';
+export * from './PersonaSubmitter.js';
